# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Furanomi API
  version: 1.0.0
  description: |
    近くのお店の空き状況をチェックできるWebアプリケーションのAPI仕様書。
    
    ## 認証について
    このAPIはJWT認証を使用します。ログインエンドポイントでHTTP-only cookieとして
    JWTトークンが発行されます。
    
    ## アプリケーション構成
    - ユーザーアプリ: 店舗検索・空き状況確認
    - 店舗管理者アプリ: 店舗情報編集・空き状況更新
    - システム管理者アプリ: 店舗・店舗管理者の管理
    
  contact:
    name: Furanomi Support
    url: https://github.com/ytskmt14/furanomi
servers:
  - url: http://localhost:3001/api
    description: 開発環境
  - url: https://api.furanomi.com/api
    description: 本番環境
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT認証トークン（HTTP-only cookie）
  schemas:
...
    ShopManager:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          example: "demo"
        email:
          type: string
          format: email
          example: "manager@example.com"
        firstName:
          type: string
          example: "太郎"
        lastName:
          type: string
          example: "山田"
        phone:
          type: string
          example: "090-1234-5678"
        is_active:
          type: boolean
          example: true
        shop:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SystemSettings:
      type: object
      properties:
        key:
          type: string
          example: "search_radius"
        value:
          type: string
          example: "5000"
        description:
          type: string
          example: "検索半径（メートル）"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid credentials"
paths:
...
  /system/settings:
    get:
      tags:
        - System
      summary: システム設定取得
      description: システム全体の設定を取得します（システム管理者のみ）
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 設定取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SystemSettings'
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 権限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - System
      summary: システム設定更新
      description: システム全体の設定を更新します（システム管理者のみ）
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - settings
              properties:
                settings:
                  type: array
                  items:
                    type: object
                    properties:
                      key:
                        type: string
                      value:
                        type: string
              example:
                settings:
                  - key: "search_radius"
                    value: "5000"
                  - key: "max_shops_display"
                    value: "50"
      responses:
        '200':
          description: 設定更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  updatedSettings:
                    type: array
                    items:
                      $ref: '#/components/schemas/SystemSettings'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 権限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /system/shop-managers:
    get:
      tags:
        - System
      summary: 店舗管理者一覧取得
      description: すべての店舗管理者の一覧を取得します（システム管理者のみ）
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 店舗管理者一覧取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopManager'
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 権限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - System
      summary: 店舗管理者作成
      description: 新しい店舗管理者を作成します（システム管理者のみ）
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - firstName
                - lastName
              properties:
                username:
                  type: string
                  example: "new_manager"
                password:
                  type: string
                  format: password
                  example: "secure_password123"
                email:
                  type: string
                  format: email
                firstName:
                  type: string
                  example: "太郎"
                lastName:
                  type: string
                  example: "山田"
                phone:
                  type: string
                  example: "090-1234-5678"
      responses:
        '201':
          description: 店舗管理者作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  manager:
                    $ref: '#/components/schemas/ShopManager'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 権限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: ユーザー名が既に使用されています
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /system/shop-managers/{id}:
    put:
      tags:
        - System
      summary: 店舗管理者更新
      description: 店舗管理者の情報を更新します（システム管理者のみ）
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 店舗管理者ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
                  description: 変更する場合のみ指定
                email:
                  type: string
                  format: email
                firstName:
                  type: string
                lastName:
                  type: string
                phone:
                  type: string
                is_active:
                  type: boolean
      responses:
        '200':
          description: 店舗管理者更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  manager:
                    $ref: '#/components/schemas/ShopManager'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 権限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 店舗管理者が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - System
      summary: 店舗管理者削除
      description: 店舗管理者を削除します（システム管理者のみ）
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 店舗管理者ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 店舗管理者削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 権限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 店舗管理者が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...